Add the travelers to a reservation
## Purpose :
Add the travelers to a reservation and add it to the database `jpa`.
Two ways are used: **servlet** and **Rest** : after the `index.html` page (user and mdp), in the `main.jsp` page, there are two choices/links, that lead to:
 * `CreateReservation-REST.jsp` form ==> **REST part**.
 * `CreateReservation-servlet.jsp` from ==> **Servlet part**.
### Servlet part:
The form calls the servlet: `AddReservationServlet`, which will act according to the clicked button.
 * if it's `Add Voyageur`, it will add the `voyageur` to `list voyageur`
 * if it's `Validate Reservation`, it will create a `reservation`, add the `list voyageur` previously created to the `reservation` and add this reservation to the database.
### REST part:
The thin client is `CreateReservation-REST.jsp`, which will lead to REST links contained in `servicesReservation.java`. And depending on the button clicked the same procedure as before will be executed.
_Ps 1 :_ I had a problem to combine Rest (`@Post`) and spring injection (`@Repository`) so `new DAOReservation` is used instead of dependency injection. So there is a duplication, the methods, noted with 1 at the end, are for the servlet part.
_Ps 2_ : `HashMap()` is used at the DAO, but I'm not sure of its position.
_Ps 3_ : I didn't finish the find part,